tools/texture_automation/requirements.txt
diffusers
transformers
accelerate
torch
numpy
Pillow
scipy
opencv-python
sentencepiece
protobuf
peft
transformers
